contract-tool/docker/requirements.txt
camunda-external-task-client-python3==4.5.0
pydantic>=1.10,<3
requests>=2.24.0